🚚 Move ajax handler to separate folder
